Add an eb-deploy script to deploy applications
This adds a script which can deploy an application to a named
environment at a specific Docker tag. We're making the following
assumptions:

1. All our applications are packaged as single-image/single-container
   applications which can be deployed to Elastic Beanstalk using a
   single JSON file as configuration.

2. The responsibility to build and publish the application images to a
   public registry lies with the application build job.

The `eb-deploy` script takes a template file -- "Dockerrun.aws.json" --
which specifies the image name and associated data (such as the port on
which the application listens) and inserts the desired tag before
publishing it as a new "application version" in Elastic Beanstalk. It
then triggers an environment update to use this new application version,
or creates the environment if necessary.