Wait for environment update to complete
It's important that we don't allow multiple deployments to the same
environment to happen at the same time. Elastic Beanstalk won't allow
this and at the moment the `eb-deploy` tool will probably crash out hard
if run while an existing environment update is already running.

The simplest way to enforce this requirement is to have the Jenkins job
configured for serial execution, and not to end the job before the
environment update is complete.

This commit adds an `eb-env-wait` script that polls the Elastic
Beanstalk API until the specified environment returns to the "Ready"
state. While doing so, it prints diagnostic events to the console, much
like the web dashboard for Elastic Beanstalk.