Add support for syncing environment settings
Unfortunately, while the environment manifest file format is very
convenient for *creating* Elastic Beanstalk environments, it's not
particularly easy to use it to *update* them.

This commit adds a script which can be used to reformat the settings
provided in the manifest file into a JSON array (which can be
interpreted by the `---option-settings` parameter to `aws
elasticbeanstalk update-environment`).

Note that this will only update or add environment option settings. It
will not remove previous settings if they are removed from the manifest.