env-wait: Abort with non-zero exit code on errors (#2)
Due to the nature of Elastic Beanstalk always trying to return back to a
ready state, builds would not fail when an error happened. This means
that developers would not get notified about a failed deployment.

This change collects all the events with a severity level of `ERROR` and
in case the `eb-env-wait` script encounters any will exit with a status
code of `1` and prints out the error messages again.