Notify Slack when deploying
This is using the `slackSend` step that is provided by the [Jenkins
Slack plugin](https://github.com/jenkinsci/slack-plugin).

First this sends a start message and then depending on the state of the
job either a success or error message to the Slack channel that is
defined in the Jenkins integration in Slack itself.