Replace 'promote' deploy type with 'redeploy'
The way that the deployment pipelines for individual applications now
work means that a separate "promote" deployment type isn't that useful.

What is very useful, however, is a deployment type that simply triggers
a redeployment of the currently deployed version to an environment, thus
cycling the infrastructure for that environment.

This commit:

- Removes the 'promote' deployment type
- Adds a 'redeploy' deployment type
- Renames the 'exact-version' deployment type to simply 'deploy' so it's
  clearer what is going on in Jenkins logs.