Add SSM command and instance ID to `eb-task-run` output
In order to view captured stdout/stderr output as a task runs, it is
necessary to know the AWS SSM command ID in order to locate the log
files in S3. Also add a message that is displayed _before_ the task
exits to tell devs where to go looking for command output.

Knowing the EC2 instance ID is also useful if one wishes to monitor a task
more closely (eg. by SSH-ing into the container/VM where it is running).