Wait for application version processing to complete
Immediately after the `create-application-version` command is executed, the
application version is in a "Processing" state because of the use of the
`--process` option. Sometimes attempts to deploy this application
version with `./bin/eb-deploy` would fail because the processing had not
yet completed by the time it ran.

Add a polling loop to the end of `./bin/eb-release` to wait until the
new application version is processed before exiting.

This should hopefully fix periodic deployment failures (eg.
https://jenkins.hypothes.is/job/deployment/1067/console).