Add Sakai to set of deployable apps
Add config to deploy Sakai image built from https://github.com/hypothesis/sakai-docker.

This will give us a persistent version of Sakai for testing the LMS app which is
available to everyone at Hypothesis.

We don't really need a QA environment for Sakai given that it is only
used for internal testing, but the Jenkinsfile script currently assumes
that both exist. We could save a few dollars by removing that
limitation.