Increase EC2 instance size for h-qa
An h-qa deploy failed due to a Docker container running out of memory
(see [1]). h-qa currently seems to use 500-700MB of memory on an
instance type that has 1GB of memory. This can be insufficient if
activity causes a spike in usage or during a deploy when a rolling
update occurs and two instances of the app are briefly running at the
same time.

Bump up the instance size to a t3a.small instead.

[1] https://hypothes-is.slack.com/archives/C074BUPEG/p1576593903011100