EB Definition - Change H-Prod instance type
The change switches the EC2 instance type being used for H-Prod.
We are migrating from M4.Large to M5.Large. The change provides a more
performant CPU at a slightly reduced cost per hour.

In addition, this update brings our configuration in-line with a
manually updated configuration made whilst troubleshooting performance
issues. The minimum number of instances we now run for H is `8`. This
maybe reduced at some point in the future when we have a better
understating of how the environment is performing under increased load.