Switch metabase to rolling deployments
As we have done for other applications, switch metabase from immutable
to rolling deployments. Rolling deployments are much faster and we've
found the trade-off vs the theoretical disadvantages to be worthwhile.

Since metabase is not an end-user facing application, I've opted to use
`Rolling` for both QA and prod rather than using
`RollingWithAdditionalBatch` for prod as we do for other services. This
trades reduced capacity during deployments for faster deployments since new
instances don't have to be spun up.