EB Definition - Move Via3 services to public subnet
This commit moves the Via3 services to our public subnet. It is a
rollback from the commit `149c6f8`.

Due to our Google Drive integration it is not possible for us to run
Via3 in the public subnets. This is because we are presenting a
consistent IP addresses when using the private subnet, and Google limits
the number of requests based on IP address.