h-qa - custom nginx
This commit delivers a custom nginx.conf for h-qa in a slightly
different manner to our norm. I have a feeling this is not supported in
the context of our Linux/Docker usage and is only supported when hosting
Java apps with Elastic Beanstalk.

The alternative to this approach will be adapting our CI to handle EB
platform hooks, but that is going to require some detailed updates.