analytics-dashboard-qa - docker nofile extension
This commit delivers a ebextension to increase the nofile softlimit
imposed by the Docker daemon. The original value was set to 1024, and
the extension raises that limit to 4096 to match the hardlimit.

In addition, updates made to the Dockerrun.aws.json file have been
removed. EB does not seem to support this method of customisation.